fix pserve to work with gevent workers
- refactored code to remove ``pyramid.threadlocal`` from pserve's import
  tree. It was coming from the ``pyramid.paster`` module which imports
  ``pyramid.scripting`` which imports ``pyramid.threadlocal``.

- dropped ``pyramid.scripts.common.logging_file_config`` as it appears
  to be unused and almost identical to ``setup_logging`` except
  without ``global_conf`` support.
